Splittet komponentene Arbeidsforholdene og Arbeidsforholdene.
