+++ src/felles-komponenter/arbeidsforholdene.less
@@ -1,21 +1,1 @@
 @import '../utils/mixins';
-
-.arbeidsforholdene {
-
-  .arbeidsforhold__enkelt {
-    .seksjonmixin();
-  }
-
-  .arbeidsforhold__detaljer {
-    .beskrivliste-mixin('row');
-  }
-
-  .arbeidsavtale {
-    padding-top:1em;
-    clear:both;
-  }
-
-  .arbeidsforhold__relevantknapp {
-    float: right;
-  }
-}
