+++ src/felles-komponenter/arbeidsforholdet.less
@@ -1,0 +1,21 @@
+@import '../utils/mixins';
+
+.arbeidsforholdet {
+
+  .arbeidsforholdet__enkelt {
+    .seksjonmixin();
+  }
+
+  .arbeidsforholdet__detaljer {
+    .beskrivliste-mixin('row');
+  }
+
+  .arbeidsavtale {
+    padding-top:1em;
+    clear:both;
+  }
+
+  .arbeidsforholdet__relevantknapp {
+    float: right;
+  }
+}
